ADD: add new validations methods Fix: fixing bug that allowed anyone to change all the data

+++ app/controllers/users_controller.ts
@@ -3,47 +3,55 @@ import { createUserValidator, updateUserValidator } from '#validators/user'
 import type { HttpContext } from '@adonisjs/core/http'
 
 export default class UsersController {
-
   async index({}: HttpContext) {
     const users = await User.query().preload('tasks')
     return users
   }
 
   async store({ request }: HttpContext) {
-    const {name, email, password} = await request.validateUsing(createUserValidator)
-    const user = await User.create({name, email, password})
-    return user  
+    const { name, email, password } = await request.validateUsing(createUserValidator)
+    const user = await User.create({ name, email, password })
+    return user
   }
 
   async show({ params, response }: HttpContext) {
-    try{
-      const user = await User.findByOrFail('id', params.id)
+    try {
+      const user = await User.findBy('id', params.id)
+      if (!user) {
+        return response.status(404).json({ error: 'User not found' })
+      }
       await user.load('tasks')
       return user
-  } catch (error){
-    return response.status(400).json({error:'User not found'})
+    } catch (error) {
+      return response.status(400).json({ error: 'User not found' })
+    }
   }
-}
 
   async update({ params, request, response }: HttpContext) {
-    try{
+    try {
       const user = await User.findBy('id', params.id)
-      const {name, password} = await request.validateUsing(updateUserValidator)
-      user!.merge({name, password})
+      if (!user) {
+        return response.status(404).json({ error: 'User not found' })
+      }
+      const { name, password } = await request.validateUsing(updateUserValidator)
+      user!.merge({ name, password })
       await user!.save()
       return user
-  } catch (error){
-    return response.status(400).json({error:'User not found'})
-  }
+    } catch (error) {
+      return response.status(400).json({ error: 'User not found' })
+    }
   }
 
   async destroy({ params, response }: HttpContext) {
-    try{
-    const user = await User.findByOrFail("id", params.id)
-    await user.delete()
-    return response.status(203)
-  } catch(error){
-    return response.status(400).json({error:'User not found'})
+    try {
+      const user = await User.findBy('id', params.id)
+      if (!user) {
+        return response.status(404).json({ error: 'User not found' })
+      }
+      await user.delete()
+      return response.status(203).json({ message: 'User deleted successfully' })
+    } catch (error) {
+      return response.status(400).json({ error: 'User not found' })
+    }
   }
-}
 }
